move form-entry-list in the same file with the form-container
