+++ src/lib/components/forms/form/form-container/terra-form-entry-list.component.html
@@ -28,21 +28,17 @@
                     [inputIsDisabled]="inputIsDisabled"
                 >
                 </terra-form-entry>
-                <ng-container *ngSwitchCase="true">
-                    <ng-container
-                        *ngTemplateOutlet="
-                            containerTemplate;
-                            context: {
-                                $implicit: inputFormField,
-                                control: control,
-                                scope: _childScopes[i],
-                                controlTypeMap: inputControlTypeMap,
-                                disabled: inputIsDisabled,
-                                width: width
-                            }
-                        "
-                    ></ng-container>
-                </ng-container>
+                <terra-form-container
+                    *ngSwitchCase="true"
+                    [formGroupName]="i"
+                    [inputFormFields]="inputFormField.children"
+                    [inputFormGroup]="control"
+                    [inputScope]="_childScopes[i]"
+                    [inputControlTypeMap]="inputControlTypeMap"
+                    [inputIsDisabled]="inputIsDisabled"
+                    [width]="width"
+                    [horizontal]="inputFormField.type === 'horizontal'"
+                ></terra-form-container>
             </ng-container>
 
             <terra-button
